+++ mex/common/primary_source/helpers.py
@@ -1,15 +1,11 @@
-from functools import cache
-
-from mex.common.models import (
-    ExtractedPrimarySource,
-)
+from mex.common.models import ExtractedPrimarySource
 from mex.common.primary_source.extract import extract_seed_primary_sources
 from mex.common.primary_source.transform import (
+    transform_seed_primary_source_to_extracted_primary_source,
     transform_seed_primary_sources_to_extracted_primary_sources,
 )
 
 
-@cache
 def get_all_extracted_primary_sources() -> dict[str, ExtractedPrimarySource]:
     """Extract and transform all primary sources.
 
@@ -28,7 +24,6 @@ def get_all_extracted_primary_sources() -> dict[str, ExtractedPrimarySource]:
     return {p.identifierInPrimarySource: p for p in extracted_primary_sources}
 
 
-@cache
 def get_extracted_primary_source_by_name(name: str) -> ExtractedPrimarySource | None:
     """Pick the extracted primary source with the given name and return it.
 
@@ -38,9 +33,9 @@ def get_extracted_primary_source_by_name(name: str) -> ExtractedPrimarySource | 
     Returns:
         Extracted primary source if it was found, else None
     """
-    try:
-        extracted_primary_source = get_all_extracted_primary_sources()[name]
-    except KeyError:
-        return None
-
-    return extracted_primary_source
+    for seed_primary_source in extract_seed_primary_sources():
+        if seed_primary_source.identifier == name:
+            return transform_seed_primary_source_to_extracted_primary_source(
+                seed_primary_source
+            )
+    return None